+++ src/components/index.ts
@@ -1,3 +1,0 @@
-// DEPRECATE: Don't know if this is really required if I'm not creating public components.
-export { default as CardItem } from './CardItem';
-export { default as Counter } from './Counter';