+++ src/containers/PoolContainer/PoolContainer.tsx
@@ -1,0 +1,12 @@
+import { Typography } from '@mui/material';
+import { type JSX } from 'react';
+
+const PoolContainer = (): JSX.Element => {
+  return (
+    <Typography variant="h6" gutterBottom>
+      Pool that contains the cards from unpacking.
+    </Typography>
+  );
+};
+
+export default PoolContainer;