+++ src/containers/index.ts
@@ -1,3 +1,0 @@
-// DEPRECATE: Don't know if this is really required if I'm not creating public components.
-export { default as HomePage } from './HomePage';
-export { default as CardList } from './CardList';